chore: move connect() protocol helper to LocalUtils
This is a local utility, does not make sense for remote connection.
